service/requirements.txt
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-Script==2.0.6
Flask==1.0.3
Jinja2==2.10.1
SQLAlchemy==1.3.4
Werkzeug==0.15.4
alembic==1.0.10
flasgger==0.9.2
itsdangerous==1.1.0
pytz==2019.1
requests==2.22.0
setproctitle==1.1.10
six==1.12.0

flake8==3.7.7
isort==4.3.20
pip-upgrader==1.4.12
pytest-cov==2.7.1
pytest-sugar==0.9.2
pytest==4.6.2
safety==1.8.5

Flask-JWT-Extended==3.18.2
Flask-Cors==3.0.7

pyemd==0.5.1
gensim
pytorch-pretrained-bert

passlib
editdistance

#torch==0.4.0
#torchvision

